+++ Netkan/Validators/InstallsFilesValidator.cs
@@ -1,6 +1,8 @@
 using System.Linq;
+
 ﻿using CKAN.NetKAN.Model;
 using CKAN.NetKAN.Services;
+using CKAN.Extensions;
 
 namespace CKAN.NetKAN.Validators
 {
@@ -30,15 +32,29 @@ namespace CKAN.NetKAN.Validators
                         mod.DescribeInstallStanzas()
                     ));
                 }
-                
+
+                // Get the files the module will install
+                var allFiles = _moduleService.FileDestinations(mod, file).Memoize();
+
                 // Make sure no paths include GameData other than at the start
-                var gamedatas = _moduleService.FileDestinations(mod, file)
+                var gamedatas = allFiles
                     .Where(p => p.StartsWith("GameData") && p.LastIndexOf("/GameData/") > 0)
                     .ToList();
                 if (gamedatas.Any())
                 {
                     var badPaths = string.Join("\r\n", gamedatas);
                     throw new Kraken($"GameData directory found within GameData:\r\n{badPaths}");
+                }
+
+                // Make sure we won't try to overwrite our own files
+                var duplicates = allFiles
+                    .GroupBy(f => f)
+                    .SelectMany(grp => grp.Skip(1))
+                    .ToList();
+                if (duplicates.Any())
+                {
+                    var badPaths = string.Join("\r\n", duplicates);
+                    throw new Kraken($"Multiple files attempted to install to:\r\n{badPaths}");
                 }
             }
         }